+++ main.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from sklearn.model_selection import train_test_split
 from Model.model import ModelHandler
-from Model.pipline import DataPipeline
+from Model.pipeline import DataPipeline
 
 def load_data():
     """
